Fix bugs found while testing file-based state tracking
This PR was intended to enable file-based state tracking. However, I found some
bugs during testing which involves touching many files, so this PR addresses
these bugs without enabling file-based state tracking.

1. I discovered that I left out a call to change the state to "running" once the
proxy-runner is ready. Simply adding the state manager type led to a circular
dependency, which required a bunch of refactoring of code. This is responsible
for most of the diff here.

2. The code which created the base directory for the status files was using the
wrong path - this was fixed.

3. I originally had separate methods for creating the initial status, and
updating. This led to problems with the detached process flow (which would try
to create the status twice, leading to an error). This PR changes the code to
use a single method which creates or update.