Ensure that file statuses reflect proxy and runtime status
With the file-based statuses, we have the potential to go out of sync
with the containers and proxy processes. Every time we read the file, if
the file's status is "running", then check the status of the workload in
the runtime, and check the proxy process. If there is disagreement, set
the status in the file to "unhealthy".

We may want to tweak this behaviour in future, for example, either
deleting the unhealthy workloads, or perhaps trying to fix them. This is
intended as a first step, since toolhive currently does not do a good
job of detecting these problems.