Consistently use maps instead of slices for env vars (#1398)
This change is intended to allow env vars to be specified to the API as a map of string to string. Some refactoring was made to apply this change consistently throughout the codebase. Now, the CLI is responsible for converting from a list of strings to the map.

Also replace a manual mock with one generated by mockgen.