Lazily register JWKS endpoint with a 5 second timeout instead of on startup
When working on patches for PRs #1470 and #1471 I've been initially
seeing very long timeouts on toolhive startup. The underlying reason
were the transports not allowing me but the fact that the timeout was so
long and happened on toolhive startup was really irritating. I think we
should fetch the JWKs when we needed, not on startup in order to not
block the startup.