Refactor tool middlewares configuration mechanism.
This change prepares the ground for the implementation of tool
override by changing the way tool middlewares are configured.

As of now, tool middlewares accept a flat list of strings acting as a
filter on the tools available to the Client. This is not sufficient
once tool override is implemented, since we have to distinguish
between a tool being filtered, being overridden, or both.

This is meant to be 100% backwards compatible, as the feature will be
implemented in a follow-up PR.