Add tool name override to tool middleware. (#1612)
This change makes use of the provided filter and override configuration
in tool middlewares. Both middlewares now process `tools/call` and
`tools/list` commands by checking both overrides and filtering,
specifically:

* `tools/list` responses, which the proxy received from the real MCP
server, must be modified to contain tools whose names might be
overridden (as well as filtered)
* `tools/call` requests, which the proxy received from the Client, must
be modified to refer to the original tool name

As a refresher for the reviewer, keep in mind that filtering is
expressed in terms of tool names used by the client, so the user-visible
(or client-visible) name must be determined before the lookup on the
filter. This makes it so that

* `tools/list` must first look the override up and only then check the
filter, while
* `tools/call` must first check the filter and only then look the
override up

Fixes #1513