Fix issue with `@latest` for npx workloads
The change introduced in #1601 introduced a bug when a user specified @latest
as the version of the npx package. The build process would create a container
with the latest version of the MCP server, but npx would fail because npx will
always check npm to see what the latest version is, and the --no-install flag
blocked it from doing so.

Do some tweaking to drop the version specifier when npx is run. This will cause
npx to just use the locally installed module that was populated during the
build.