Add methods to status manager for storing PIDs (#1719)
This is part of a process of replacing the dedicated pid files with the
status files. This is needed for k8s since we do not need PID files, and
writing them creates problems.

This change does not alter how PID files are managed: it just provides
the operations needed to add them to status files, and implements tests.