Extend the CRD with a way to set proxyrunner's pod template spec annotations
Adds a new CRD field `PodTemplateMetadataOverrides`. This field allows
to set metadata (labels, annotations) on the pod template of the proxy
runner workload - this is different from us being able to set the
metadata on the MCP server itself through the pod template field in the
MCPServer CRD as well as the existing metadata overrides which only work
for the deployment itself.