Refactor `RunConfigBuilder` to Option Pattern.
This change replaces the builder pattern used to build a `RunConfig`
with the `OptionPattern`. I haven't added any additional validation
inside the various `WithXXX` routines, but this work can be further
improved moving some of the logic written in
[validateConfig](https://github.com/stacklok/toolhive/blob/1aba1802c56494052770532f1408de07d20341b6/pkg/runner/config_builder.go#L573)
out of it, further simplifying run config validation procedures.

The resulting encapsulation is visible in `runConfigBuilder` now being
no longer accessible outside the package.