Stop Stop calling PID file functions from pid.go in tests
Some of the tests directly manipulate PID files. In the case of the e2e
test - change the tests to use workload statuses. In file_status_go -
we need to test PID file migration, so include simplified logic for
reading PID files into the test so that the functions in process/pid.go
can be made private.