Add support for Tool Override to CLI.
This change adds a new `--tools-override` option to `thv run` and
similar commands. The new option accepts a path to a file containing a
JSON structure like the following

```json
{
  "toolsOverride": {
    "actual-name": {
      "name": "new-name",
      "description": "Overridden description."
    }
  }
}
```

The override file is read once when the MCP server is started, and is
not reloaded on `thv restart`. This behaviour might change in the
future.

Note that Tool Override feature is mostly orthogonal to Tool Filtering
one, with the exception of the name used for filtering.

Specifically, overrides just change the name or description of an
existing tool, but do not limit access to other tools; filtering does
the exact opposite. When both an override and a filter are specified
for the same tool, the name to be used when specifying the filter is
the overridden name. We decided to implement it this way for two
reasons

1. the user is likely to think of tools in the MCP server in terms of
   the new names rather than the actual ones, since the former are
   those visible to the Client, and
2. the MCP spec does not specify means for MCP servers to publish the
   full list of tools without first running it, and determining those
   available requires starting the server itself and issuing a
   `tools/list` call, i.e. require acting as a client.

As a consequence, it's impossible to reliably determine the full list
of tools, and a valid filter in one scenario might not work in another
one.

Fixes #1511