use default naming for workloads when name is not provided
When we do not provide the workload name it is unable to verify whether
the workload already exists causing re running the same workload. This
change make sure we use the default consistent naming and even if the
name is not provided we can check whether the workload is already running.