Force reload of config singleton after writing config changes
We ran into a bug with the `SetupComplete` flag and the UI. It turned
out that the config was being updated, but the singleton was not updated
so the API treated the flag as false even after it had been updated. Fix
this by forcing the config save code to reload after a successful write.

In future, we may want to move mutable flags like this one out of the
config and into a more suitable home.