Add `pkg/testkit` package
This change adds a new `testkit` package that should help us write e2e
in a simpler way. Its sole purpose is

* adding utilities to quickly spin-up an HTTP test server exposing
  either a Streamable-HTTP or (legacy) SSE MCP server
* adding utilities to ease the parsing of `text/event-stream` response
  bodies

The file `pkg/testkit/testkit_test.go` contains a few tests that
exemplify how to use the framework, which is barebone at the
moment. Ideally, it should allow the developer to add assertions in
the test server as well, but for now it only allows configuring the
returned JSON payloads.