Fix handling of `text/event-stream` during authz
This change fixes a bug of the authorization middleware that prevents
it from filtering tools list if the content type is something other
than `application/json`.

The original code implicitly assumes that the response's content type
is `application/json`, while it can be `text/event-stream` in both
Streamable-HTTP and SSE setups. Solving that exposed another issue
with the way JSON RPC 2.0 responses were parsed by the same
middleware.

Fixes #1189