Set correct content length when mutating calls (#2132)
This sets content length correctly when `tool/call` requests are
modified by the Tool Config middleware. The issue manifests only when
`pkg/transport/proxy/transparent` is used, which makes it a bit
trickier to test.

I manually tested the fix, but I've yet to find a decent way to test
it.

Fixes #2056