Remove registry-related labels from the example config
I must have modified the example fetch server when working on the
registry feature with enforcing registry labels and commited the diff by
mistake.

Once we promoted the registry enforcement feature out of experimental,
this started hitting us because the MCPServer suddendly started
requiring that a registry exists.

Let's remove the labels.