Fix server port when network mode is `host`
Docker networking works differently depending on whether it's running
inside a VM or on the (linux) host itself, with VM-based installations
like Docker Desktop not supporting `host` network. What happens is
that the port is exposed on the VM rather than the actual host. Note
that this is the case for linux running Docker Desktop.

On the other hand, Docker Engine installations work just fine on
linux, but we were still configuring the Proxy to try to connect to
the wrong port.

This is PR assumes that
* all linux installations run docker engine
* all non-linux installations run on VMs

Fixes #2317