Decouple transport from runtime
As currently implemented, `pkg/transport` is tightly coupled with the
underlying runtime. Specifically, transport implementations are aware
of the fact that they're wrapping a container workload and even use
`rt.Deployer` to stop and restart the workload. This coupling makes it
hard to test the transport layer.

This is the first of a series of patches that aim to decouple the
transport layer from the underlying workload.

First of a series of PRs addressing #2429