Do not convert volume target path to Windows style
Since we deal with Linux containers, the volume target path should
always be a Unix style path, even on Windows. Prior to this, the code
would convert the target path to a Windows style path on Windows hosts
due to the use of filepath.Clean. This PR changes the code to use
path.Clean for the target instead, which always produces a Unix style
path. Pre-existing validation ensures that the user cannot pass Windows
style paths as input.