removes MCPRegistry datasources from Operator (#2568)
* removes storage manager from datasources package

as part of the moving of code from the operator to the registry api, the
storage configmap will be refactored later on into mounting the registry
json ONLY if it's specified as a configmap.

Signed-off-by: ChrisJBurns <29541485+ChrisJBurns@users.noreply.github.com>

* removes data sources

Signed-off-by: ChrisJBurns <29541485+ChrisJBurns@users.noreply.github.com>

* removes call to unneeded method and disable test temporarily

Signed-off-by: ChrisJBurns <29541485+ChrisJBurns@users.noreply.github.com>

* removes unused code

I've had to modify some existing code in order to satisfy lint and
tests. some of this will get put back in a later PR

Signed-off-by: ChrisJBurns <29541485+ChrisJBurns@users.noreply.github.com>

* temporarily disables tests until we add the `--config` code

Signed-off-by: ChrisJBurns <29541485+ChrisJBurns@users.noreply.github.com>

---------

Signed-off-by: ChrisJBurns <29541485+ChrisJBurns@users.noreply.github.com>