moves e2e ext auth test to int test (#2625)
* moves e2e ext auth test to int test

we don't need the heavy e2e tests anymore and the integration tests are
a quick and fast way of asserting the correct things are being created.
I've added a couple of integration tests here that test a bit more than
the e2e tests (finalisers, updates etc) however during writing of the
int tests it found a bug that the e2e tests let slip through. it seems
that the referenced servers aren't being put into the status of the
externalauth CR. I have commented this test/assertion out for now so we
can use it later when we submit the fix to the controller.  I will make
a separate ticket for this

Signed-off-by: Chris Burns <29541485+ChrisJBurns@users.noreply.github.com>

* merging with main

Signed-off-by: Chris Burns <29541485+ChrisJBurns@users.noreply.github.com>

* removes unused function

Signed-off-by: Chris Burns <29541485+ChrisJBurns@users.noreply.github.com>

* oops wrong assertions... damnn claude

Signed-off-by: Chris Burns <29541485+ChrisJBurns@users.noreply.github.com>

---------

Signed-off-by: Chris Burns <29541485+ChrisJBurns@users.noreply.github.com>