removes pulling and loading of registry api image in operator deploy task (#2667)
removes pulling and loading of registry api image

this was probably a breadcrumb left over from when the registry api was
in this repo. so it made sense to load it into the kind node because it
is common loading locally built images into the kind cluster for
testing. but not the registry server lives in its own repo and is
publicly avaialable, when the operator creates the deployment for the
registry server inside of the cluster it just pulls it itself. this
removing the need for us to pull the image ourselves and loading it into
kind.

there were also problems doing the pulling and loading of a remote image
because docker pulls only the image that matches your host architecture
whereas kind loads the image with `--all-platforms` as the default. this
causes errors because it can't load the full imagine manifests because
docker never pulled them for the other platforms. we could tell kind to
load it only for the platform of the host but its pointless given its a
public image now and we aren't building it as part of this task. the
operator and proxyrunner images loaded fine because we build them using
`ko` so all platforms are built by default resulting in kind having no
issues loading them.

Signed-off-by: Chris Burns <29541485+ChrisJBurns@users.noreply.github.com>