Change pinging logic for remote workloads (#2872)
Previously, we did not include the pinger for remote workloads. The
assumption was that we may not be able to authenticate while pinging the
remote, so we would not get the expected status. There were two issues
which arose:

- It turns out the pinger accepts any non 50x, so we do not need to
  worry about 401.
- By excluding the pinger from the remote workload, we lack the ability
  to test basic connectivity to the remote workload. This has led to
  some reported issues where network connectivity is lost, and ToolHive
  is not aware that there is a problem.

This PR adds the pinger to the remote workload. In future, we may want
to be more selective - for example, only allow 401s for remote
workloads.