fixes generation of manifests via kubebuilder
currently when you run a task operator-manifests, it will change the
clusterrole when it shouldn't. the reason for this is because a var
block was added directly below the kubebuilder comment markers that
generate the permissions for the clusterrole. kubebuilder therefore
ignored the markers because they thought the comments were related to
the var block itself. this commit moves the var block to be away from
the kubebuilder markers to remove interference

Signed-off-by: Chris Burns <29541485+ChrisJBurns@users.noreply.github.com>