adds reusable Kubernetes configmap functionality in Operato for C/R/U actions (#2997)
In order to reduce amount of duplication around common Kubernetes configmap functionality, this PR adds a kubernetes/configmaps package that allows you to retrieve a configmap, its values and upsert them. The motivation behind this PR is not only reduce duplication but to offer common helper methods to be reused throughout the operator now and in future.

---------

Signed-off-by: Chris Burns <29541485+ChrisJBurns@users.noreply.github.com>